Move store and load diff functions back to commitment factory
